Fix Vulkan/Asahi Linux compatibility: Handle CPU backend properly
- Skip torch.cuda.current_stream() on Vulkan platform
- Make torch.Event() optional (None on Vulkan/CPU)
- Fix sampled_token_ids_cpu attribute creation for Vulkan
- Remove async_copy_ready_event assertions for Vulkan
- Add platform checks before CUDA-specific operations

This allows vLLM to run on Asahi Linux with Vulkan backend
and CPU fallback when device memory is limited.

Tested: Chat completion requests now succeed on M1 Max.

Fixes:
- AssertionError: Torch not compiled with CUDA enabled
- RuntimeError: cpu backend doesn't support events
- AssertionError: async_copy_ready_event is not None

+++ vllm/v1/worker/gpu/block_table.py
@@ -74,7 +74,7 @@ class BlockTables:
         self.slot_mappings = torch.zeros(
             self.num_kv_cache_groups,
             self.max_num_batched_tokens,
-            dtype=torch.int64,
+            dtype=torch.int32,  # VULKAN: int64 not supported
             device=self.device,
         )
